+++ example/lib/screens/basic_example_screen.dart
@@ -43,6 +43,8 @@ class _BasicExampleState extends State<BasicExample> {
                     )
                   : null,
             ),
+            if (_currentPhoto != null) const SizedBox(height: 15),
+            if (_currentPhoto != null) Text("Author: ${_currentPhoto.author}"),
             const SizedBox(height: 15),
             RaisedButton(
               child: const Text("Fetch photo"),